requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
supabase>=2.0.0
asyncpg>=0.29.0

# Task Queue
redis>=5.0.0
celery>=5.3.0

# Web Scraping & Automation
scrapy>=2.11.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytrends>=4.9.0
playwright>=1.40.0

# Messaging
python-telegram-bot>=20.7
telethon>=1.33.0

# AI & ML
openai>=1.3.0
google-generativeai>=0.3.0
Pillow>=10.0.0

# Utilities
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Google Analytics 4
google-analytics-data>=0.18.0
google-analytics-admin>=0.26.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Notifications
apprise>=1.6.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Captcha Solving
2captcha-python>=1.2.0

